finished higherlower

+++ schema.sql
@@ -38,7 +38,7 @@ CREATE TABLE IF NOT EXISTS movies (
     genre_id INT NOT NULL,
     overview VARCHAR(2048) NOT NULL,
     poster_img_url VARCHAR(255) NOT NULL,
-    vote_average INT NOT NULL,
+    vote_average FLOAT NOT NULL,
     createdAt BIGINT NOT NULL
 );
